Fix stray null byte in Makefile
